Updating Documentation for Hashing
Bret Mulvey, the author of the article cited in this pulication has migrated his work to papa.bretmulvey.com. I was able to view an archival version of Bret M.'s original post (http://home.comcast.net/~bretm/hash/3.html) and have validated that this is the same paper that is originally cited.